python-syncer/pyproject.toml
[tool.poetry]
name = "syncer"
version = "0.1.0"
description = ""
authors = ["ilaif <ifallach@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
schedule = "^1.2.0"
psycopg2-binary = "^2.9.6"
pygithub = "^1.59.0"
sqlalchemy = "^2.0.17"
pydantic = {version = "^1.10", extras = ["dotenv"]}
loguru = "^0.7.0"
alembic = "^1.11.1"
atlassian-python-api = "^3.40.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
